+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - rendering statics for user with membership
 - rendering members of statics
 - search engine
+- api endpoint characters#search
 
 ## [0.6.0] - 2019-11-01
 ### Added